Make it possible to run more tests in parallel
